+++ rlimit_other.go
@@ -1,9 +1,9 @@
-// +build !darwin
+// +build !darwin,!freebsd
 
 package main
 
 import "syscall"
 
 func rlimitMax(rlimit syscall.Rlimit) uint64 {
-	return uint64(rlimit.Max)
+	return rlimit.Max
 }
